+++ test/bigtest/tests/jobLogs-test.js
@@ -9,6 +9,10 @@ import commonTranslations from '@folio/stripes-data-transfer-components/translat
 import translations from '../../../translations/ui-data-export/en';
 import { setupApplication } from '../helpers';
 import { jobLogsContainerInteractor } from '../interactors';
+import {
+  testUserOzzy,
+  testUserElliot,
+} from '../network/scenarios/fetch-job-profiles-success';
 
 const getCellContent = (row, cell) => jobLogsContainerInteractor.logsList.rows(row).cells(cell).content;
 
@@ -25,25 +29,29 @@ describe('Job logs list', () => {
     });
 
     it('should add status column to the end', () => {
-      expect(jobLogsContainerInteractor.logsList.headers(4).text).to.equal(translations.status);
+      expect(jobLogsContainerInteractor.logsList.headers(5).text).to.equal(translations.status);
     });
 
     it('should be sorted by "completedDate" descending by default', () => {
-      expect(getCellContent(0, 4)).to.equal('Fail');
-      expect(getCellContent(1, 4)).to.equal('Success');
+      expect(getCellContent(0, 5)).to.equal('Fail');
+      expect(getCellContent(1, 5)).to.equal('Success');
     });
 
     it('should render ID column', () => {
       expect(jobLogsContainerInteractor.logsList.headers(1).text).to.equal(commonTranslations.jobExecutionHrId);
     });
 
+    it('should render Run by column', () => {
+      expect(jobLogsContainerInteractor.logsList.headers(4).text).to.equal(commonTranslations.runBy);
+    });
+
     describe('clicking on status header', () => {
       beforeEach(async () => {
-        await jobLogsContainerInteractor.logsList.headers(4).click();
+        await jobLogsContainerInteractor.logsList.headers(5).click();
       });
 
       it('should sort by status in ascending order', () => {
-        expect(getCellContent(1, 4)).to.equal('Success');
+        expect(getCellContent(1, 5)).to.equal('Success');
       });
 
       it('should have the correct query in path', function () {
@@ -52,12 +60,12 @@ describe('Job logs list', () => {
 
       describe('clicking on status header', () => {
         beforeEach(async () => {
-          await jobLogsContainerInteractor.logsList.headers(4).click();
+          await jobLogsContainerInteractor.logsList.headers(5).click();
         });
 
         it('should sort by status in descending order', () => {
-          expect(getCellContent(1, 4)).to.equal('Fail');
-          expect(getCellContent(0, 4)).to.equal('Success');
+          expect(getCellContent(1, 5)).to.equal('Fail');
+          expect(getCellContent(0, 5)).to.equal('Success');
         });
 
         it('should have the correct query in path', function () {
@@ -88,6 +96,24 @@ describe('Job logs list', () => {
 
       it('should have the correct query in path', function () {
         expect(this.location.search.includes('direction=ascending&sort=hrId')).to.be.true;
+      });
+    });
+
+    describe('clicking on Run by header', () => {
+      beforeEach(async () => {
+        await jobLogsContainerInteractor.logsList.headers(4).click();
+      });
+
+      it('should sort by User name from Run by field', () => {
+        const userOzzy = `${testUserOzzy.firstName} ${testUserOzzy.lastName}`;
+        const userElliot = `${testUserElliot.firstName} ${testUserElliot.lastName}`;
+
+        expect(getCellContent(0, 4)).to.equal(userElliot);
+        expect(getCellContent(1, 4)).to.equal(userOzzy);
+      });
+
+      it('should have the correct query in path', function () {
+        expect(this.location.search.includes('direction=ascending&sort=runBy')).to.be.true;
       });
     });
   });